[chore]: update coverage
